% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riceCH_eg.R
\docType{data}
\name{riceCH_eg}
\alias{riceCH_eg}
\title{Rice canopy height example data}
\format{
\describe{
  \item{date}{Date of observation}
  \item{height}{CH}
  \item{id}{Cultivar ID}
  \item{x}{Dayes after sowing}
}
}
\usage{
riceCH_eg
}
\description{
This data was a simulated rice CH data to show the usage of this package.
}
\keyword{datasets}
